1. Добавлены классы для представления таблиц proff.prof_competence и proff.competence_model; 2. Синхронизировано состояние БД с состоянием БД в ветке DB; 3. Правки типов данных некоторых сущностей; 4. Рефакторинг (косметические правки кода, реструктуризация пакетов, связанных с новыми таблицами); 5. Мелкие доработки; 6. Включение режима validate в пропертях spring boot

+++ src/main/java/ru/prostandard/model/dicts/EducationLevel.java
@@ -10,10 +10,10 @@ public class EducationLevel implements Serializable {
     @Id
     @Column(name = "level_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "level_value")
-    private Long levelValue;
+    private Integer levelValue;
 
     @Column(name = "description")
     private String levelDescription;
@@ -22,25 +22,25 @@ public class EducationLevel implements Serializable {
 
     }
 
-    public EducationLevel(Long id, Long levelValue, String levelDescription) {
+    public EducationLevel(Integer id, Integer levelValue, String levelDescription) {
         this.id = id;
         this.levelValue = levelValue;
         this.levelDescription = levelDescription;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
-    public Long getLevelValue() {
+    public Integer getLevelValue() {
         return levelValue;
     }
 
-    public void setLevelValue(Long levelValue) {
+    public void setLevelValue(Integer levelValue) {
         this.levelValue = levelValue;
     }
